(Level changes)
The level "ConnorsLevel," now has spikes .

The general basis of this level is that the camera doesn't move at all, the whole level is inside of the initial camera view. The player will see two buttons, one that they can get to, and one they can not. Upon pressing the first button, the door in front of the second button will open and platforms will appear and allow the player to jump to the button. Upon pressing the second button, the door in front of the goal will be opened and platforms will appear and allow the player to get to the end of the level