Spikes
I am requesting the addition of a new level: It's contains a new empty for hazards  which in turn contains a spikes empty which also contains several of the new spikes prefab.
I have also created a new script folder for hazards which I have added a new hurt script that makes it so hazards can hurt the player.